Allow wrapping a static reference in an Rc or Arc
This is a counterproductive thing to do in isolation since it just uses
three pointers worth of space to store a single pointer, but it can be
useful when storing a collection of pointers where some are reference
counted while others are static.

This is effectively type erasure of the difference between an Rc<T>/Arc<T>
and a &'static T, and comes at no additional storage cost because the base
types from the standard library have a niche bit pattern that can represent
the absense of an allocation with no additional storage.